feat(offers):Добавлены новые модели Country,Region,City

+++ backend/offers/models/advertisement.py
@@ -27,6 +27,8 @@ class Country(models.Model):
 
     def __str__(self):
         return self.name
+
+
 class Region(models.Model):
     """Регион"""
 
@@ -48,6 +50,7 @@ class Region(models.Model):
     def __str__(self):
         return self.name
 
+
 class City(models.Model):
     """Город"""
 
@@ -67,10 +70,6 @@ class City(models.Model):
 
     def __str__(self):
         return self.name
-
-
-
-
 
 
 class Advertisement(
@@ -122,7 +121,7 @@ class Advertisement(
 
     category = models.CharField("Категории", max_length=100, choices=CategoryChoices.choices)
     title = models.CharField("Название", max_length=100)
-    price = models.PositiveIntegerField("Цена", null=True, blank=True)
+    price = models.DecimalField("Цена", max_digits=6, decimal_places=2,  null=True, blank=True)
     coordinates = PlainLocationField(verbose_name="Координаты", blank=True)
     description = models.TextField("Описание", max_length=1000, null=True, blank=True)
     is_published = models.BooleanField("Опубликованно", default=True)
@@ -170,4 +169,3 @@ class AdvertisementImage(models.Model):
 
     def __str__(self):
         return f"Объявление - {self.advertisement.title}, #{self.pk}"
-
